+++ src/main/kotlin/dev/dediamondpro/resourcify/gui/injections/ImageButton.kt
@@ -15,7 +15,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package dev.dediamondpro.resourcify.gui.pack
+package dev.dediamondpro.resourcify.gui.injections
 
 import com.mojang.blaze3d.systems.RenderSystem
 import net.minecraft.client.gui.GuiGraphics