+++ changelog.md
@@ -1,5 +1,10 @@
 ## Resourcify 1.7.1
 
+Please make sure to report any bugs and/or visual anomalies
+on [Resourcify's GitHub](https://github.com/DeDiamondPro/Resourcify/issues) or in
+the [Discord](https://discord.gg/XtAuqsJWby).
+
+- Move downloading worlds to "Create new world" and then press the "Download" tab.
 - Add updating support for Datapacks when using a mod
   like [Datapack/Gamerule Menus](https://modrinth.com/mod/packrule-menus).
 - After updating a Resource Pack, you are now returned to the pack selection GUI, instead of immediately triggering a